Execute better external processes

+++ commands/docker.js
@@ -1,7 +1,11 @@
-const exec = require('../lib/exec');
+const proc = require('../lib/process');
 const fs = require('fs-extra');
 const S = require('string');
 const path = require('path');
+const Promise = require('bluebird');
+const yargonaut = require('yargonaut');
+const chalk = yargonaut.chalk();
+
 module.exports = function(yargs) {
 
   yargs
@@ -21,27 +25,48 @@ module.exports = function(yargs) {
 
         Promise.resolve()
           .then(function() {
+            return fs.ensureDir('tmp');
+          })
+          .then(function() {
+
+            if (!fs.existsSync(pathPkgLock)) {
+              return;
+            }
+
+            return fs.readFile(pathPkgLock)
+              .then(function(content) {
+                if (!content.length) {
+                  return fs.remove(pathPkgLock);
+                }
+              });
+
+          })
+          .then(function() {
 
             if (!fs.existsSync(pathPkgLock)) {
               console.log(`Init empty ${pathPkgLock}`);
               return fs.writeFile(pathPkgLock, '{}');
             }
+
           })
           .then(function() {
             // Init empty cache file
             if (!fs.existsSync(CACHE_NAME)) {
 
               console.log(`Init empty ${CACHE_NAME}`);
-              return exec.run(`tar cvzf ${CACHE_NAME} --files-from /dev/null`);
+              return proc.exec(`tar cvzf ${CACHE_NAME} --files-from /dev/null`);
 
             }
 
           })
           .then(function() {
-            return exec.run('docker-compose build dev');
+            return proc.spawn('docker-compose build dev');
           })
           .then(function() {
-            return exec.run(`docker run --rm --entrypoint cat ${CONTAINER_NAME}:latest /tmp/package-lock.json > tmp/package-lock.json`);
+            console.log('Extracting package-lock.json');
+            return proc.exec(`docker run --rm --entrypoint cat ${CONTAINER_NAME}:latest /tmp/package-lock.json > tmp/package-lock.json`,{
+              silent: true
+            });
           })
           .then(function() {
 
@@ -55,29 +80,46 @@ module.exports = function(yargs) {
             if (source != target) {
               console.log('Saving NPM cache');
 
-              exec.run(`docker run --rm --entrypoint tar ${CONTAINER_NAME}:latest czf - /.cache/npm/ > ${CACHE_NAME}`)
+              proc.exec(`docker run --rm --entrypoint tar ${CONTAINER_NAME}:latest czf - /.cache/npm/ > ${CACHE_NAME}`,{
+                silent: true
+              })
                 .then(function() {
                   console.log('Saving package-lock.json');
-
+                  if (!fs.existsSync(pathPkgLockTmp)) {
+                    return;
+                  }
                   return fs.copy(pathPkgLockTmp, pathPkgLock);
 
                 });
             }
 
-          });
+          })
+          .then(function() {
 
+            //run cleanup commands so you don't amass images that you don't need
+            console.log('Cleanup images');
+            return proc.exec('docker rm -v $(docker ps -a -q -f status=exited) 2>&1')
+              .then(function() {
+                return proc.exec('docker rmi $(docker images -f "dangling=true" -q) 2>&1');
+              });
+
+          })
+          .then(function() {
+            console.log();
+            console.log(chalk.bold.green('Success!'));
+          });
       });
 
       yargs.command('run', 'Run the portal', function() {
 
-        exec.run('docker-compose up -d dev')
+        proc.exec('docker-compose up -d dev')
           .then(function() {
 
             var terminal = 'x-terminal-emulator -e "docker-compose exec dev bash -c';
 
-            exec.run(`${terminal} 'tail -n 1000 -f .pm2/logs/output-0.log'"`);
-            exec.run(`${terminal} 'tail -n 1000 -f .pm2/logs/error-0.log'"`);
-            exec.run(`${terminal} 'pm2 list; exec /bin/bash -i'"`);
+            proc.exec(`${terminal} 'tail -n 1000 -f .pm2/logs/output-0.log'"`);
+            proc.exec(`${terminal} 'tail -n 1000 -f .pm2/logs/error-0.log'"`);
+            proc.exec(`${terminal} 'pm2 list; exec /bin/bash -i'"`);
           });
 
       });
